Issues solved with the affective cycle and run functions modified
Issues solved with the affective cycle.
The run functions of the environment and AffectiveAgent have also been modified for this purpose.
Now the run function of the environment invokes the run function of the agents. The run function of each agent is in charge of executing the affective and rational cycle of the agent.