Updated Affective Categories and Example
The behavior of the affective category annotations in the plans has been modified.

Now it is enough that the agent contains one of the affective categories annotated in the plan (not all of them) for the plan to be executed.

The example with affective categories has also been updated to show this behavior in a clearer way.